gateway/requirements.txt
awsiotsdk==1.22.0
pyserial==3.5
